Renamed graph mod to async_detect mod
